requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
Pillow==10.1.0
feedparser==6.0.10
python-dotenv==1.0.0
pydantic==2.5.0
SQLAlchemy==2.0.23
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
asyncio==3.4.3
playwright==1.40.0
lxml==4.9.3
pytesseract==0.3.10
openai==1.3.7
anthropic==0.7.0
youtube-dl==2021.12.17
yt-dlp==2023.11.16
imagehash==4.3.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
sentry-sdk==1.38.0
loguru==0.7.2
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.2
tenacity==8.2.3
jsonschema==4.20.0
python-slugify==8.0.1
unidecode==1.3.7
langdetect==1.0.9
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
opencv-python==4.8.1.78
moviepy==1.0.3
ffmpeg-python==0.2.0
